requirements.txt
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.4.0
pytest-cov>=4.1.0

